requirements.txt
copernicusmarine
h5py
python-dotenv